+++ src/E5R.Architecture.Core/TransformationManager.cs
@@ -51,14 +51,6 @@ namespace E5R.Architecture.Core
 
             return to;
         }
-
-        public TTo Transform<TFrom, TTo, TOperation>(TFrom from, TOperation operation)
-            where TTo : new() where TOperation : Enum
-        {
-            Checker.NotNullArgument(from, nameof(from));
-
-            return ResolveType<ITransformer<TFrom, TTo, TOperation>>().Transform(@from, operation);
-        }
         #endregion
 
         #region List transformer
@@ -93,15 +85,6 @@ namespace E5R.Architecture.Core
                 return to;
             });
         }
-
-        IEnumerable<TTo> ITransformationManager.Transform<TFrom, TTo, TOperation>(IEnumerable<TFrom> from, TOperation operation)
-        {
-            Checker.NotNullArgument(from, nameof(from));
-
-            var t = ResolveType<ITransformer<TFrom, TTo>>();
-
-            return from.ToList().ConvertAll(t.Transform);
-        }
         #endregion
 
         #region Paginated list transformer
@@ -119,15 +102,6 @@ namespace E5R.Architecture.Core
             Checker.NotNullArgument(from, nameof(from));
 
             var result = AutoTransform<TFrom, TTo>(from.Result);
-
-            return new PaginatedResult<TTo>(result, from.Offset, from.Limit, from.Total);
-        }
-
-        PaginatedResult<TTo> ITransformationManager.Transform<TFrom, TTo, TOperation>(PaginatedResult<TFrom> from, TOperation operation)
-        {
-            Checker.NotNullArgument(from, nameof(from));
-
-            var result = ((ITransformationManager)this).Transform<TFrom, TTo, TOperation>(from.Result, operation);
 
             return new PaginatedResult<TTo>(result, from.Offset, from.Limit, from.Total);
         }